Tax authorities

+++ Projects/Application/Interfaces/Persistence/IBankRepository.cs
@@ -4,6 +4,6 @@ namespace NUCA.Projects.Application.Interfaces.Persistence
 {
     public interface IBankRepository : IRepository<MainBank>
     {
-        public Task<bool> BankHasProjects(int id);
+        public Task<bool> BankHasProjects(long id);
     }
 }
